# Conflicts:
#	PaymentService/src/main/kotlin/it/polito/wa2/lab5/g09/paymentservice/utils/TransactionDeserializer.kt

+++ TicketCatalogueService/src/main/kotlin/it/polito/wa2/lab5/group09/ticketcatalogueservice/entities/TicketCatalogue.kt
@@ -12,7 +12,7 @@ data class TicketCatalogue (
     @Indexed
     val ticketId: ObjectId = ObjectId.get(),
     var type: String,
-    var price: Float,
+    var price: Double,
     var zones : String,
     val maxAge : Int? = null,
     val minAge : Int? = null,
